requirements.txt
django==3.2.22
django-environ==0.11.2
django-cors-headers==4.2.0
djangorestframework==3.14.0
Pillow==10.0.1
django-storages==1.14.2
django-ckeditor==6.7.0
psycopg2==2.9.9